+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- ...
+- Renamed `fiboa create` to `fiboa create-geoparquet`
 
 ### Deprecated